Tiny refactors
